agent: platform_manager: add timer factory and timer manager
https://jira.prplfoundation.org/browse/PPM-776

Add timer manager member variable in platform_manager. Create
timer factory and then timer manager in main and pass it to the class
constructor.

This is a preparatory commit. Next commits will use provided timer
manager to create timer instances whenever it is required to execute
actions on a regular basis or after a given timeout has elapsed

Signed-off-by: Mario Maz <mmazzapater@gmail.com>