agent: Filter AP-DISABLED notification
There are scenarios in which HOSTAPD could send AP DISABLED notification
as part of an internal configuration transition.
For example when we use RECONF or switch to a DFS channel,
the HOSTAPD will disable the radio and we get an "AP-DISABLED" notification.
This is unwanted behaviour from the Agent perspective since when the
son_slave get AP-DISABLED notification, we restart and disconnect the
ap_manager (and kill its process).

To avoid such behaviour, add a filter on the AP-DISABLED Ap_manager
and Monitor event handlers which wait 3 seconds (magic number),
and if the state of the radio has changed from DISABLED so higher state,
then the DISABLED notification is not being forwarded to the son_slave.

Also, when hostapd is disabled refresh radio info can fail
so don’t break and keep waiting until timeout expires.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>
Signed-off-by: Alex Kanter <alex.kanter@devalore.com>