bcl: fix netlink events are not fired
Netlink events are not being fired by the Netlink event listener
implementation. The reason is that the buffer being used to hold
received Netlink messages is a member variable, so it is reused on
each call and gets full after a few messages are received.

The solution has been to use a local variable instead of a member
variable so the buffer is empty on each usage.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>