controller: Fix authentication type check
The bss authentication type is checked against the supported
authentication types of the agent upon sending the M2. The current
check is a simple AND operation with not equal to 0. This check is
incorrect since it can miss some conditions, i.e:
m1->auth_type_flags = 0x20 (WPA2)
bss_info_conf.authentication_type = 0x60 (WPA2+SAE)
The result would be non zero and the INFO will not be printed.

Fix the condition to check that all the bits of the requested BSS remain
on.

Signed-off-by: Semyon Barenboym <semyon.barenboym@devalore.com>