dev: boardfarm: check_status using beerocks cli moved to base class
_run_shell_cmd function signature:
pexpect.run has a inner validation checking if args is an instance of type
 list, which would make an exception if we ever decided not to fill this
 argument.

Removing env from pexpect.run call:
Pexpect run function already defaults the env param to None, so
making a pre-condition to call it with and without the env is
unnecessary redundancy.

Added a get_active_entity function:
Both compose/prplwrt classes share the self.role attribute, we can have a
get_active_entity function on prplmesh_base and get it from inheritance in
order to avoid constantly checking whether a device is supposed to be an agent
or a controller. Thought it was a good idea to make a function that just
returns the one that was chosen on instantiation, since there can/should only
be one anyway.

Use of UCCSocket:
check_status now uses the UCCSocket class from capi.py (ucc_socket.cmd_reply)
to make a beerocks call.
Since now we're using ucc_socket to send the command, we can also move
check_status to the main class (prplmesh_base) and avoid duplicated code.

Signed-off-by: David <davidmgafonso@gmail.com>