gitlab-ci: retry jobs that fail occasionally
Some of our CI jobs fail occasionally.

The rdk-b build fails regularly because it fails to update the git
clones from github. Since the clones are cached, there is no real need
to do the update, but unfortunately bitbake still fails even though the
required commit is already there. This is always a transient failure, so
a retry should fix it (although there's a risk it will fail on a
different repository in the next build).

test_flows fails occasionally on various tests. Since this is being
migrated to boardfarm, we don't want to bother debugging it in
test_flows. So just retry.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>