db: fix filtering clients when max db size reached
When the loading clients from the persistent DB and
the max persistent DB size limit is reached, it
is required to filter out the clients with the least
remaining.
Fix the calculation of how many clients should be
filtered out due to the above-mentioned limit.
Change the for-each filtering loop direction from
`max-size to end` instead of `end to max-size`.
Fix the filtered clients log-print (was printing the
negative diff value instead of 0 which means no clients
are filtered due to max-db-size limit reached).

Signed-off-by: Daniel Edri <daniel.edri@devalore.com>