gitlab-ci: split run-certification-tests between docker and platform
The run-certification-tests jobs need the prplmesh binary that was built
in an earlier stage as an artifact. Depending on the DEVICE_UNDER_TEST
setting, this can be either the Ubuntu binaries or one of the platform
binaries.

With the child pipeline, we solved this with a complicated `needs`
statement that pulls in the artifacts from jobs in the child pipeline.
We also added a dependency on the `platform` job, to be sure that that
job is finished before starting the run-certification job.

However, the result is that if any job in the child pipeline fails, the
`platform` job will fail, and the `run-certification-tests` job won't
even start. Although obviously no such failures should occur on master,
sometimes they do (case in point: RDK-B build failed because it couldn't
contact github to download one of its packages).

We could simply remove the dependency on the `platform` job, but then we
would *always* use the artifacts from a different run of the pipeline.
This is, at best, confusing.

As an alternative, duplicate the run-certification-tests jobs: one for
the docker version (DEVICE_UNDER_TEST == "prplmesh"), and one for the
platform version (DEVICE_UNDER_TEST != "prplmesh"). The latter one is
moved into the platform child pipeline. This way, we only have a
dependency on jobs in earlier stages of the same pipeline. The
dependency on the `platform` job can be removed.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>