agent: backhaul_manager: add UCC server
https://jira.prplfoundation.org/browse/PPM-753

Add UCCServer member variable in backhaul_manager_thread to use it when
agent_ucc_listener is created. Create UCCServer in main and pass it to
the class constructor.

This is a preparatory commit. Next commits will use provided UCC
server to exchange certification commands and replies with a client
instead of current TCP server socket. For now, the line that creates
the UCC server is commented out to avoid having two TCP server sockets
listening on the same port.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>