ci: add a stability test for boardfarm
From time to time issues sneak into master because they don't make
tests fail on every run, but only from time to time.

Add a "stability" job that will be run when the "$STABILITY_TESTS"
variable is defined. It will run the boardfarm tests 10 times in the
same pipeline.

This will allow us to run this boardfarm stability test in a separate
nightly job, which could alert us whenever _anything_ is failing in
the job (because if it was included in another scheduled job, it would
be more difficult to notice that it failed).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>