bcl: create CMDU client interface and implementation
https://jira.prplfoundation.org/browse/PPM-814

Create a client component to exchange CMDU messages with a CMDU server.
This component depends on a socket connection established with a server
socket, on a parser to read CMDU messages from a socket, and on a
serializer to write CMDU messages to a socket.

The CmduClientImpl class extends the CmduPeer class. Next commit will
modify CmduServerImpl class to extend CmduPeer too.

In a future MR, an instance of CmduClient will be used from backhaul
manager to connect to the CmduServer running in the platform manager
and exchange CMDU messages.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>