beerocks_defines: add configuration files path
Preparative Commit.
In some platforms, the path where the prplmesh is installed is
read-only, and to provide a way to edit configuration files for
debugging the configuration files are copied to a writeable path.
The above means we need the prplmesh to check if the configuration file
exists on that path before trying to use the read-only, local, relative
path.

Add CONF_FILES_PATH that is determined at compile-time and is
platform-specific.

PPM-789.

Signed-off-by: Adam Dov <adamx.dov@intel.com>