gitlab-ci: rename DEVICE_UNDER_TEST for individual certification tests
The DEVICE_UNDER_TEST variable is used both for run-certification-tests
and for the individual certification test jobs. For
run-certification-test, it is supposed to be set by the triggerer
(schedule or web), while for the individual jobs, they are set by the
job itself.

However, if this variable is set by the triggerer, it will override the
value that is set in the yaml file for the individual certification test
jobs. So these jobs don't work properly any more.

Similarly, now the DEVICE_UNDER_TEST variable is passed on from the
parent to the child pipeline, it will *always* override the value that
is set in the individual job, even if DEVICE_UNDER_TEST is not set.

To fix this, rename the variable to RUN_ON_DEVICE for the individual
jobs. This way, it is no longer overridden by the triggerer or the
parent pipeline.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>