bcl: fix bug in UCC command parser
https://jira.prplfoundation.org/browse/PPM-790

Fix bug in parser implementation when parsed command contains white
spaces. If this is the case, then not all parsed bytes are removed from
the buffer with received data (because we are shifting the lenght of
command after trimming instead of the whole frame, including white
spaces).

Also add unit tests to verify that the bug has been fixed.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>