agent: Choose the second best channel
After receiving the channel list, the ZWDFS flow should select the second best
channel and then send it to the ZWDFS FH to switch antenna from 2.4G radio
and switch to this channel if it a DFS channel.

Add to the CHANNELS_LIST_RESPONSE handler a switching to the
ZWDFS CHOOSE_NEXT_BEST_CHANNEL state.
Add to CHOOSE_NEXT_BEST_CHANNEL ZWDFS state code that selects the
second best channel and request from the ZWDFS FH to switch to this
channel and start CAC.
If the selected channel is the failsafe channel (the channel the radio is currently at),
then the ZWDFS flow is aborted.
If the selected channel is not a DFS channel, then the state machine skips several
states and immediately moves to the state SWITCH_CHANNEL_PRIMARY_RADIO.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>