agent: Fix decision to switch ZWDFS antenna
When iterating on the channel bandwidth sub 20MHz channels on
the zwdfs_select_best_usable_channel() function to check whether any
of the sub beacon channels is unavailable so we would know to not
choose the channel, we need also set the DFS state of the channel with
the higher bandwidth to the worst DFS state among of its overlapping
beacon channels.
For example, if some channel at bw=80MHz has two beacon channels
which are not DFS, one AVAILABLE channel, and one USABLE channel,
the DFS state of that channel should be USABLE.
Set the DFS state by the worst DFS state, and add a jump to
SWITCH_CHANNEL_PRIMARY_RADIO state if the DFS state of the channel
is AVAILABLE since this DFS state does not require CAC similar to the
NOT_DFS state.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>