agent: Fix handler of BACKHAUL_ZWDFS_RADIO_DETECTED
Two issues are fixed here:
1. Add increment if the iterator of the loop.
It was missing and caused the handler to stuck.
2. Change the getter function of the front interface name to the getter which returns char*.
The getter which returns std::string construct the string with the defined size of the string.
The problem is that if the string is defined with a fixed size, the constructed string will have
this size, which means garbage characters after the printable characters.
If the return type is char* any string constructed from it will discard the non-printable
characters.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>