agent: Filter AP-DISABLED notification
Apparently, when we command the HOSAPD to switch to a DFS channel,
the HOSTAPD disabled the radio, and we get an "AP-DISABLED" notification.
When the CAC is completed, we get back "AP-ENABLED".
This is unwanted behavior from the Agent perspective since when the
son_slave get AP-DISABLED notification, we restart and disconnect the
ap_manager (and kill its process).

To avoid such behavior, add a filter on the AP-DISABLED Ap_manager
and Monitor event handlers which wait 3 seconds (magic number),
and if the state of the radio has changed to DFS, then the DISABLED
notification is not being forwarded to the son_slave.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>