agent: Add/Update debug prints
Add debug print in several places.
On the son_slave, remove debug print about receiving messages from
ap_manager when the master socket is nullptr.
Since the condition is only printing messages without actually doing
something, remove it.
The main reason for it is that on ZWDFS radio, there is no link to the
controller so it floods the log.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>