agent: Switch off the ZWDFS antenna on failures
Currently, when failures occur on ZWDFS flow, the FSM moves to state "NOT_RUNNING".
It is not optimal error handling because on some states the ZWDFS antenna is switched
and moving to the "NOT_RUNNING" state will leave the 2.4G radio short with once antenna.
To optimize that, when an error occurs after the ZWDFS has switched the antenna, move
to state "ZWDFS_SWITCH_ANT_OFF_REQUEST" instead.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>