nbapi: add unit test for Ambiorix class
For now, the test only verifies the initialisation/deinitialisation and
a single variant of the set() function.

For the initialisation/deinitialisation, it uses a stubbed backend
(inspired on the tests within ambiorix itself).

Since the set() function doesn't end up calling anything in the backend,
it is verified by checking the value after the set. Since we don't have
a getter function in the Ambiorix abstraction at the moment, that is
done directly with the amxd functions.

The odl file is defined in the source code and written to a temporary
file. It includes all possible types already, for use in later tests.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>