gitlab-ci: set ENABLE_NBAPI in build-in-docker
For our tests, we want to start using the north-bound API. So we need to
build with NBAPI enabled in CI.

Only the builder-docker-ubuntu image currently has the ambiorix
libraries, so we can't simply enable it in build-in-docker in general.
Therefore, turn build-in-docker into a template, and instantiate it with
ENABLE_NBAPI=ON in extra cmake flags. Make sure that the name doesn't
change, so we don't need to update dependencies.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>