controller: add infrastructure for dynamic_channel_selection_r2_task
This commit is the first in a series of commits introducing the
main flow and related messages for the dynamic channel selection r2 task,
running under the controller.

The dynamic channel selection r2 task will receive requests for
channel scans, communicating with the channel scan task in the agent,
and manage the scans by keeping track of the agents and radios status.

This commit adds infrastructure for the task.
- Add TRIGGER_SCAN state to FSM.
- Add structs for scan_request and scan_report events.

PPM-360
---
Signed-off-by: Oren Vormaser <oren.vormaser@devalore.com>