ci: build for nec-wx3000hp (scheduled jobs only)
Since we don't need it in our regular CI, adding it to the scheduled
jobs will makes sure it doesn't slow down the regular pipelines, while
still allowing us to know from day to day if it still builds.

Other platform builds are in a child pipeline, but in child pipeline
we can't use CI_PIPELINE_SOURCE to run only on scheduled
pipelines (because it will be set to 'parent_pipeline').
Because of this the new job is added to the main pipeline (it will be
moved whenever we start using the NEC device in MR merge requests).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>