ci: reboot all the devices before nightly tests
The issue in https://jira.prplfoundation.org/browse/PPM-798 where one
of the interfaces of the axepoint was not working anymore was gone
after a reboot of the device.

Always reboot all the devices before nightly tests run.

Update the "easymesh_cert" repo before the reboot to make sure the
correct version of power_cyle.py is used.

Also improve the logs when deploying the firmware/prplMesh by echoing
what is going to be done.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>