bcl: extend serialize_cmdu() to allow forwarding
https://jira.prplfoundation.org/browse/PPM-826

This is a preparatory commit.

In a follow-up commit we are going to include a method to forward a
CMDU message received from the broker server running in the transport
process to another local process.

Among other information, the broker server includes the index of the
interface the CMDU message was received on (in the UDS header).
However, currently the method to serialize a CMDU message does not
allow to specify the interface index.

Thus, modify the `serialize_cmdu()` method to include the interface
index the message was received on. Update unit tests accordingly.

Also, since `serialize_cmdu` is going to be used to serialize both
CmduMessageTx and CmduMessageTx classes, modify the signature to accept
common base class CmduMessage. Move the `CmduMessageTx::finalize()`
call out of the serialization method to `CmduPeer::send_cmdu()`, just
before serializing the CmduMessageTx to send.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>