agent: backhaul_manager: add routing params to forward CMDU
https://jira.prplfoundation.org/browse/PPM-753

This is a preparatory commit in order to eventually replace current
CMDU forwarding mechanism with the new `CmduServer::forward_cmdu()`
method.

When forwarding a received CMDU message from backhaul manager to a
slave thread, the same UDS header received together with the CMDU shall
be forwarded too. Currently this is possible only because UDS header is
in the same buffer as the CMDU data. However, the new `forward_cmdu()`
method does not include such restriction and UDS information must be
provided as parameters to the method.

So add parameters required to fill in the UDS header of a forwarded
CMDU message to method `backhaul_manager::forward_cmdu_to_uds`.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>