btl: prepare to add forwarding method
https://jira.prplfoundation.org/browse/PPM-820

This is a preparatory commit.

In a follow-up commit we are going to include a method to forward a
CmduMessageRx received from a slave through a UDS socket to the broker
server running in the transport process to another local process.

Currently we have a method to send a CmduM820ageTx to the broker server
that shares some code with the new method to be implemented.

To avoid having duplicate code, refactor existing method and move out
shared code into a new private method that will also be called from the
forwarding method to be implemented.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>