db: prefer removal of aging clients over non-aging clients
In case we want to add a new client to the persistent DB but the DB
reached his max-size - we need to remove a client according to the
following preference (ascending):
- prefer to keep in the DB non-aging clients (clients with internal
time-life configured to 0).
- prefer to keep in the DB connected clients.
- prefer to keep in the DB the clients with the most remaining
time-life.

The non-aging configuration which was recently added required changing
the way we find the candidate client to remove from the DB to make-room
for the new client configuration.

Adding clients to persistent DB when DB is full will prefer the removal
of aging clients over non-aging ones.

PPM-808.

Signed-off-by: Adam Dov <adamx.dov@intel.com>