controller: db: Too many error logs from NBAPI stuff
After running prplmesh on the UGW platform
Moran Shoeg found the problem - when prplmesh
built with AmbiorixDummy class data model is not
create and methods which tries to get the correct
path to the different data model objects generates
error logs, in some cases a lot of error logs.

Fix is simple - cover the methods for getting
path to the object by ENABLE_NBAPI flag and if
flag turned off - use the dummy implementation of
those methods.

Signed-off-by: Vladyslav Tupikin <v.tupikin@inango-systems.com>