agent: backhaul_manager: add method to send CMDU to broker
https://jira.prplfoundation.org/browse/PPM-753

This is a preparatory commit in order to eventually replace current
mechanism to send CMDU messages from backhaul manager to the broker
server. Currently, CMDU messages are sent to broker server using a
method in base class `btl::transport_socket_thread`. In a follow-up
commit, `backhaul_manager` will not extend this class any more and
CMDU messages will be sent using the new
`BrokerClient::send_cmdu()` method.

Create method `backaul_manager::send_cmdu_to_broker_temp()` to send a
CMDU to broker server using current mechanism. The code in this method
will later be replaced with a call to `BrokerClient::send_cmdu()`. The
name of the method includes a "_temp" suffix because there is already a
method with the same name in base class. This suffix will later be
removed, when `backahul_manager` will no longer extend that class.

Next commit will change backhaul manager to use this new method
whenever a CMDU is sent to broker server for dispatching.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>