boardfarm: capi_wireless_onboarding: add missing sniffer start call
The capi_wireless_onboarding.py was missing the call to start the
sniffer, so it was consistently failing with:
get_packet_capture but no capture file

Because the sniffer returned an empty result without raising an
exception and because an empty result is a valid case for
"check_no_cmdu_type", the test was reported as "passing" for
boardfarm.

To make sure it doesn't happen again in the future, make the sniffer
raise an exception instead of printing an error and returning an empty
result when something went wrong. This will make sure the tests fail
when there is an issue with the sniffer, like checking for a cmdu
before the sniffer was started.

Once all tests have been ported to boardfarm, we'll be able to use a
boardfarm exception for those errors. For now use ValueError.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>