btl: add function create_broker_client_factory
https://jira.prplfoundation.org/browse/PPM-753

Currently we have code to create an instance of a broker client factory
(a factory to create broker clients connected to broker server running
in the transport process) in beerocks_master_main.cpp. In next commits,
this code will also have to be added to beerocks_slave_main.cpp.

To avoid having duplicated code, move such code to a new common factory
method where it can be reused.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>