controller: use function create_broker_client_factory
https://jira.prplfoundation.org/browse/PPM-753

Refactor code in beerocks_master_main.cpp to use the new function to
create an instance of a broker client factory.

While we are at it, change the type of smart pointer used by the
controller to hold the reference to the broker client factory from
shared_ptr to unique_ptr (the ownership of the factory is transferred
to the controller).

Signed-off-by: Mario Maz <mmazzapater@gmail.com>