agent: platform_manager: init file descriptors for timers.
There are two member variables in platform manager class holding the
file descriptors for the timers used in the class. These timers are
optional so they might never get to be created. Since they are not
initialized to the `invalid_value` value, the `stop()` method tries
to remove them on exit and an error message is logged.

Initialize the file descriptors of the timers so the timers get removed
on exit only if they have been created.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>