boardfarm: prplmesh_prplwrt: start prplMesh before creating the entity
Because creating the ALEntity requires the UCC listener to be running,
we need to start prplMesh before it.

The following changes are introduced:

- If the mode is controller, we also need to start prplMesh. Replace
'prplmesh_start_agent' with 'prplmesh_start_mode', which takes care of
starting prplMesh in the right mode and optionally wait for some delay
to let prplMesh the time to initialize.

- Instead of returning a value, make prplmesh_start_mode raise an
exception if we didn't get the prompt back.

- Add "delay" to the configuration file and use it for
"prplmesh_start_mode". Its initial value is 30 seconds.

- Improve the docstring.

Note that if we also used a generic 'entity', instead of having
separate controller_entity and agent_entity, we could also get rid of
the 'if' condition (having 2 separate 'entity' members is redundant
since each entity has 'is_controller'). This also require changes to
prplmesh_compose however, so it's not done here.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>