btl: add method to configure transport process
https://jira.prplfoundation.org/browse/PPM-821

Include a method in broker client to build and send the configuration
message to the broker server running in the transport process. The
configuration message includes the name of the network bridge to use in
the transport process.

In a future MR we will use this new method in the backhaul manager, to
set the network bridge to use in the transport process. This method
will replace existing `configure_ieee1905_transport_interfaces()`
defined in transport_socket_thread, which is deprecated in favor of the
use of the event loop and a single threaded application.

Note that since we added the BridgeStateManager in the transport
process, the configuration message does not need a list of interfaces
but only the name of the bridge.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>