bcl: group all broker client handlers in a struct
https://jira.prplfoundation.org/browse/PPM-815

Broker client defines two event handlers to deal with on connection
closed and on CMDU received events. Since the two handlers are always
set together, it would be convenient to have a single function to set
all of them at once, similar to how it is done in the EventLoop class.

This is the first of a series of commits to accomplish that task:
create a struct that groups all handlers and add the methods to set
and clear the handlers all at once. In a follow-up commit these new
methods will be used instead of the old ones.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>