boardfarm: add iface_to_device
To run on an actual device, boardfarm needs to know which interface is
connected to it.

Add "iface_to_device" to the configuration file. The interface is then
added to the bridge which contains the docker controller, to allow
them to communicate.

Note that add_iface_to_bridge was previously used to change the bridge
settings on the prplWrt device. However, The bridge settings of the
prplWrt device is supposed to be handled by the device, through the
UCI setting.  Change add_iface_to_bridge to add_host_iface_to_bridge
to make it clear that it applies to the host interfaces and bridge.

The value of iface_to_device is currently hard-coded to the one that
will be used in CI. A future task may change it to make it easier to
use different values.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>