boardfarm: rax40, wifi: change default values
Change the default value of the interface name of the wifi device to
the one that will be used by the CI runner.

The wan_ip for the rax40 didn't have an initial value in the config
file, so add one.

A future task may introduce changes to make it easier to use different
values.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>