tools: boardfarm: move the packages needed for the tests to the end
The packages that we need for the tests are likely to change more
often the the ones to install docker and docker-compose. Having them
installed at the top of the Dockerfile means we're likely to have to
rebuild the image from the start.

Un-comment the line that installs the dependencies needed by the
tests, and remove tshark and wireshark-common from the first call to
apt.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>