boardfarm: run initial_ap_config on the netgear rax40
Use the changes to dctest.py and boardfarm to run the first test on
the rax40.

A new boardfarm test suite is added to allow us to separate the tests
we run in docker and the tests we run on device, since not all of them
will work as is on the device.

A detail that makes this commit more complicated than it should be, is
caused by the fact that the boardfarm job in the child pipeline needs
the docker build from the parent pipeline.

A first attempts was made by using a similar approach than for the
'upload-artifacts' jobs: use a 'needs' to get artifacts from the
parent pipeline. However, this turned out to always fail with the
following message:

This job depends on other jobs with expired/erased artifacts:
Please refer to
https://docs.gitlab.com/ee/ci/yaml/README.html#dependencies

When checking for corresponding job on Gitlab, it was always there and
built before the platform pipeline started. Possibly moving the
boardfarm test to the main pipeline would work, but that in turn would
bring us back to the situation where we cannot add merge requests to
the merge train before the expensive builds are triggered (because the
boardfarm tests also depends on those expensive builds).

Another explanation to this error could be that the 'ref' attribute
currently doesn't support branch names with forward slashes (see
below). For the 'upload-artifacts' job, this has never been a problem
since it only runs on the 'master' branch.

Since using Gitlab's mechanism to download the artifacts automatically
didn't work, the next approach that was considered was to use Gitlab's
API to manually download the latest artifacts available for the
'build-in-docker' job, on the branch that is currently tested.  In
theory, it should have been as simple as doing:

curl --location --output artifacts.zip --header "JOB-TOKEN: $CI_JOB_TOKEN" "$CI_API_V4_URL/projects/$CI_PROJECT_ID/jobs/artifacts/$CI_COMMIT_REF_NAME/download?job=$JOB_NAME"

However, when used on a development branch, this always returned a
'404 Not found'. Since the same command works for the 'master' branch,
a possible explanation is that it fails because development branches
include forward slashes, and this specific Gitlab endpoint does not
support it.

The following was then tested:

- Try another endpoint to see if URL-encoded branch names are
supported at all. First getting information about the branch *without*
URL-encoding the branch name. It fails with a 404 error:
https://gitlab.com/api/v4/projects/20535651/repository/branches/dev/PPM-660-boardfarm-add-device-support

- Getting information about the branch by URL-encoding the branch
name. It works:
https://gitlab.com/api/v4/projects/20535651/repository/branches/dev%2FPPM-660-boardfarm-add-device-support

- Finally, downloading the artifacts for 'build-in-docker' for the
development branch by URL-encoding the branch name in the same way. It fails with a
404 error:
https://gitlab.com/api/v4/projects/20535651/jobs/artifacts/dev%2FPPM-660-boardfarm-add-device-support/download?job=build-in-docker

- Note that the same call for the master branch worked:
https://gitlab.com/api/v4/projects/20535651/jobs/artifacts/master/download?job=build-in-docker

Finally, since using the API to directly download the artifacts from
the 'ref' didn't work, another approach (also using the API, but
different endpoints) was used. First, use the parent pipeline
ID (explicitly exported by the parent pipeline) to get a list of every
jobs for the pipeline, and extract the job ID of
'build-in-docker'. From there, to use the Gitlab API again to download
the artifacts for that specific build ID.

The following changes are introduced:

tests:

- Add a new test suite: 'test_flows_on_prplwrt'. Every time we can
pass a new test on the device, it will be added there.

ci/templates.yml:

- .test-on-target: Move deploying firmware and ipk to the
'before_script', to be able to re-use it for boardfarm.

- Since the boardfarm test will switch the device to "agent" mode, we
need to make sure it's set back to controller mode before the
operational test starts, so add it in the beginning of
'test-on-target'. Note that there will be a period of time where this
branch and the master branch run this test on the device, while at the
same time other branches won't have this change. To try to not break
the existing tests for side branches until they are rebased on master,
the controller mode is also set back at the end of the boardfarm
test (see below).

ci/platform.yml:

- Add boardfarm-on-netgear-rax40. It extends .test-on-target to
benefit from before_script, then runs the boardfarm
'test_flows_on_prplwrt' test suite.

- The boardfarm test needs a prplMesh controller running in docker, so
use Gitlab API to get the job ID of 'build-in-docker' from the parent
pipeline, then manually download its artifacts (see above for a
lengthier explanation about why this is needed). Using 'JOB-TOKEN'
with '$CI_JOB_TOKEN' gives us '{"message":"403 Forbidden"}', so use a
separate private token created only for that instead. The private
token only has read access to the api, repository, and registry.

- The device we have in CI does not use 'eth1' for its wired backhaul,
but eth0_1. At the moment the only way to set it is by modifying
beerocks_agent.conf. Change it manually, with a TODO note to configure
it in UCI once it's possible (i.e. once PPM-230 is done).

- After the tests, reset the target to "controller" mode. While this
is not strictly necessary, it's an attempt to not break the existing
tests for branches which do not have this change yet.

.gitlab-ci.yml:

- Since the boardfarm job needs the docker build (for the docker
controller) the platform stage is moved after the 'build' stage, and
the 'platform' job has 'build-in-docker' added to its 'needs' to make
sure it doesn't start before the docker builds are done. It also needs
the boardfarm-ci image, so add a 'need' for it as well.

- The child pipeline needs the parent pipeline ID, so set it as a
variable.

Note that the new boardfarm job uses the same tags as the operational
tests, to allow us to keep the latter running on the Turris Omnia and
GL.iNet.

The following resources help to understand how downloading the
artifacts work:
https://docs.gitlab.com/ee/api/jobs.html
https://docs.gitlab.com/ee/api/job_artifacts.html
https://docs.gitlab.com/ee/api/pipelines.html

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>