bcl: ucc_listener: wait for reset completed in a thread
https://jira.prplfoundation.org/browse/PPM-753

Currently backhaul manager and UCC listener are two different threads
that run in parallel. So there is no problem if UCC listener sits in a
loop waiting for the backhaul manager to set a boolean value meaning
that "reset is completed" and then continue.

After completing this task, both backhaul manager and UCC listener will
run in main thread and will wait for blocking I/O operations on the
application event loop. Thus the above scenario will not be an option.

Therefore, to return from the reset command only when reset is
complete, create a thread exclusively dedicated to perform the wait and
detach it.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>