boardfarm: allow running from child pipelines
When running from a child pipeline, CI_PIPELINE_ID will refer to the
child pipeline. However, since the docker images were built in the
parent pipeline, there will be no image in the registry with the child
pipeline ID.

tools: docker-compose.yml:

- Change CI_PIPELINE_ID to IMAGE_ID. We anyway didn't use the variable
directly, since when running locally it was set in 'dctest.py'.

dctest.py:

- If PARENT_PIPELINE_ID is set, we're running from a child pipeline so
use it as IMAGE_ID. If not, check if CI_PIPELINE_ID is set. If it is,
then we're running in CI, but from the main pipeline. Else it means
we're running locally, so use the "latest" image.

PARENT_PIPELINE_ID is currently not set anywhere, but it will be set
in a later commit when the parent pipeline triggers the child. Note
that PARENT_PIPELINE_ID doesn't start with 'CI_', first to avoid
confusing it with Gitlab's predefined variables, but also because it's
not unlikely that Gitlab would use that variable name in the future.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>